Update Unit test to accept arguments fot addition testing

+++ AdditionLogic/Addition.cs
@@ -4,7 +4,7 @@ namespace AdditionLogic
 {
     public static class Addition
     {
-        public static double Add(int firstNum, int secondNum)
+        public static double Add(double firstNum, double secondNum)
         {
             return firstNum + secondNum;
         }
